Move globalEnvs logic to service editor
